requirements.txt
asgiref==3.6.0
Django==4.1.7
django-autoslug==1.9.8
django-ckeditor==6.5.1
django-extensions==3.2.1
django-extra-fields==3.0.2
django-js-asset==2.0.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
order==2.1.1
Pillow==9.4.0
product==0.1.1.1
PyJWT==2.6.0
python-slugify==8.0.1
pytz==2022.7.1
scinum==1.4.4
six==1.16.0
sqlparse==0.4.3
text-unidecode==1.3
tzdata==2022.7